requirements.txt
comtypes==1.4.13
Pillow==12.0.0
pycaw==20251023
pynput==1.8.1
pystray==0.19.5
